data_iep/requirements.txt
fastapi
uvicorn[standard]
qdrant-client
mysql-connector-python
google-api-python-client
google-auth
tqdm
python-dotenv
